fix queue

+++ src/main/java/it/gov/pagopa/fdrxmltojson/QueueTriggerFn.java
@@ -5,12 +5,10 @@ import com.azure.core.util.Context;
 import com.azure.data.tables.TableClient;
 import com.azure.data.tables.models.ListEntitiesOptions;
 import com.azure.data.tables.models.TableEntity;
-import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.datatype.jsr310.JavaTimeModule;
 import com.microsoft.azure.functions.ExecutionContext;
 import com.microsoft.azure.functions.annotation.BindingName;
-import com.microsoft.azure.functions.annotation.ExponentialBackoffRetry;
 import com.microsoft.azure.functions.annotation.FunctionName;
 import com.microsoft.azure.functions.annotation.QueueTrigger;
 import it.gov.pagopa.fdrxmltojson.model.BlobData;
@@ -18,12 +16,7 @@ import it.gov.pagopa.fdrxmltojson.model.QueueMessage;
 import it.gov.pagopa.fdrxmltojson.util.StorageAccountUtil;
 
 import java.time.Duration;
-import java.time.Instant;
-import java.time.LocalDate;
-import java.time.temporal.ChronoUnit;
-import java.time.temporal.TemporalUnit;
 import java.util.Optional;
-import java.util.TimeZone;
 import java.util.logging.Level;
 import java.util.logging.Logger;
 
@@ -31,8 +24,6 @@ import static it.gov.pagopa.fdrxmltojson.util.XMLUtil.messageFormat;
 import static java.time.temporal.ChronoUnit.SECONDS;
 
 public class QueueTriggerFn {
-
-    private static final String MAX_RETRY_COUNT = System.getenv("ADD_PAYMENT_REQUEST_PARTITION_SIZE"); // it should be maxDequeueCount + 1
 
     private static Logger logger;
 
@@ -70,35 +61,42 @@ public class QueueTriggerFn {
 
         QueueMessage queueMessage = objectMapper.readValue(message, QueueMessage.class);
 
-        BlobData blobData = StorageAccountUtil.getBlobContent(queueMessage.getFileName());
+        if (queueMessage != null &&
+                !queueMessage.getContainerName().isBlank() && queueMessage.getContainerName().equals(StorageAccountUtil.getFdr1FlowBlobContainerName()) &&
+                !queueMessage.getFileName().isBlank()
+        ) {
+            BlobData blobData = StorageAccountUtil.getBlobContent(queueMessage.getFileName());
 
-        String sessionId = Optional.ofNullable(blobData.getMetadata().get("sessionId")).orElse("NA");
-        try {
-            fdrXmlCommon.convertXmlToJson(context, sessionId, blobData.getContent(), blobData.getFileName(), dequeueCount);
+            String sessionId = Optional.ofNullable(blobData.getMetadata().get("sessionId")).orElse("NA");
+            try {
+                fdrXmlCommon.convertXmlToJson(context, sessionId, blobData.getContent(), blobData.getFileName(), dequeueCount);
 
-            // if convertion is executed properly then removes the entry from fdr1conversion error table
-            TableClient tableClient = StorageAccountUtil.getTableClient();
-            ListEntitiesOptions listEntitiesOptions = new ListEntitiesOptions()
-                    .setFilter(String.format("RowKey eq '%s'", sessionId));
-            PagedIterable<TableEntity> entries = tableClient.listEntities(listEntitiesOptions, Duration.of(3, SECONDS), new Context("invocationId", context.getInvocationId()));
-            if (entries != null) {
-                for (TableEntity entry : entries) {
-                    tableClient.deleteEntity(entry);
+                // if convertion is executed properly then removes the entry from fdr1conversion error table
+                TableClient tableClient = StorageAccountUtil.getTableClient();
+                ListEntitiesOptions listEntitiesOptions = new ListEntitiesOptions()
+                        .setFilter(String.format("RowKey eq '%s'", sessionId));
+                PagedIterable<TableEntity> entries = tableClient.listEntities(listEntitiesOptions, Duration.of(3, SECONDS), new Context("invocationId", context.getInvocationId()));
+                if (entries != null) {
+                    for (TableEntity entry : entries) {
+                        tableClient.deleteEntity(entry);
+                    }
                 }
-            }
 
-        } catch (Exception e) {
-            String formattedMessage = messageFormat(sessionId, context.getInvocationId(), "NA", blobData.getFileName(), "%s error [retry attempt: %s]", operation, dequeueCount);
-            logger.log(Level.SEVERE, formattedMessage, e);
-
-            if (dequeueCount ==  Long.parseLong(MAX_RETRY_COUNT) - 1) {
-                formattedMessage = messageFormat(sessionId, context.getInvocationId(), "NA", blobData.getFileName(), "[ALERT][FdrXmlToJson][LAST_RETRY] Performed last retry for blob processing");
+            } catch (Exception e) {
+                String formattedMessage = messageFormat(sessionId, context.getInvocationId(), "NA", blobData.getFileName(), "%s error [retry attempt: %s]", operation, dequeueCount);
                 logger.log(Level.SEVERE, formattedMessage, e);
+
+                if (dequeueCount ==  getMaxRetryCount() - 1) {
+                    formattedMessage = messageFormat(sessionId, context.getInvocationId(), "NA", blobData.getFileName(), "[ALERT][FdrXmlToJson][LAST_RETRY] Performed last retry for blob processing");
+                    logger.log(Level.SEVERE, formattedMessage, e);
+                }
+
+                throw new Exception(formattedMessage, e);
             }
-
-            throw new Exception(formattedMessage, e);
         }
-
     }
 
+    private Long getMaxRetryCount() {
+        return Long.parseLong(System.getenv("MAX_RETRY_COUNT")); // it should be maxDequeueCount + 1
+    }
 }
